# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CINELOOP API - TMDB proxy for movie/TV content
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TmdbMediaItem {
  id: number;
  title?: string;
  name?: string;
  original_title?: string;
  original_name?: string;
  overview?: string;
  poster_path?: string | null;
  backdrop_path?: string | null;
  media_type?: string;
  genre_ids?: number[];
  release_date?: string;
  first_air_date?: string;
  vote_average?: number;
  vote_count?: number;
  popularity?: number;
  adult?: boolean;
}

export interface TmdbPagedResponse {
  page: number;
  results: TmdbMediaItem[];
  total_pages: number;
  total_results: number;
}

export interface TmdbVideo {
  id: string;
  key: string;
  name: string;
  site: string;
  type: string;
  official?: boolean;
}

export interface TmdbVideosResponse {
  id: number;
  results: TmdbVideo[];
}

export interface TmdbGenre {
  id: number;
  name: string;
}

export interface TmdbGenresResponse {
  genres: TmdbGenre[];
}

export interface TmdbMediaDetail {
  id: number;
  title?: string;
  name?: string;
  overview?: string;
  poster_path?: string | null;
  backdrop_path?: string | null;
  genres?: TmdbGenre[];
  release_date?: string;
  first_air_date?: string;
  vote_average?: number;
  vote_count?: number;
  runtime?: number;
  number_of_seasons?: number;
  number_of_episodes?: number;
  status?: string;
  tagline?: string;
  homepage?: string;
}

export type GetTrendingMoviesParams = {
  page?: number;
};

export type GetTrendingAllParams = {
  page?: number;
};

export type GetTrendingTvParams = {
  page?: number;
};

export type GetAnimeParams = {
  page?: number;
};

export type SearchTmdbParams = {
  query: string;
  page?: number;
};

export type GetTmdbVideosParams = {
  id: number;
  type: GetTmdbVideosType;
};

export type GetTmdbVideosType =
  (typeof GetTmdbVideosType)[keyof typeof GetTmdbVideosType];

export const GetTmdbVideosType = {
  movie: "movie",
  tv: "tv",
} as const;

export type DiscoverTmdbParams = {
  type: DiscoverTmdbType;
  genre?: number;
  page?: number;
};

export type DiscoverTmdbType =
  (typeof DiscoverTmdbType)[keyof typeof DiscoverTmdbType];

export const DiscoverTmdbType = {
  movie: "movie",
  tv: "tv",
} as const;

export type GetTmdbGenresParams = {
  type: GetTmdbGenresType;
};

export type GetTmdbGenresType =
  (typeof GetTmdbGenresType)[keyof typeof GetTmdbGenresType];

export const GetTmdbGenresType = {
  movie: "movie",
  tv: "tv",
} as const;

export type GetTmdbDetailsParams = {
  id: number;
  type: GetTmdbDetailsType;
};

export type GetTmdbDetailsType =
  (typeof GetTmdbDetailsType)[keyof typeof GetTmdbDetailsType];

export const GetTmdbDetailsType = {
  movie: "movie",
  tv: "tv",
} as const;