# lib/api-zod/src/generated/types/tmdbGenre.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CINELOOP API - TMDB proxy for movie/TV content
 * OpenAPI spec version: 0.1.0
 */

export interface TmdbGenre {
  id: number;
  name: string;
}