# lib/api-zod/src/generated/types/tmdbPagedResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CINELOOP API - TMDB proxy for movie/TV content
 * OpenAPI spec version: 0.1.0
 */
import type { TmdbMediaItem } from "./tmdbMediaItem";

export interface TmdbPagedResponse {
  page: number;
  results: TmdbMediaItem[];
  total_pages: number;
  total_results: number;
}