# lib/api-zod/src/generated/types/tmdbVideo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CINELOOP API - TMDB proxy for movie/TV content
 * OpenAPI spec version: 0.1.0
 */

export interface TmdbVideo {
  id: string;
  key: string;
  name: string;
  site: string;
  type: string;
  official?: boolean;
}