# lib/api-zod/src/generated/types/tmdbVideosResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CINELOOP API - TMDB proxy for movie/TV content
 * OpenAPI spec version: 0.1.0
 */
import type { TmdbVideo } from "./tmdbVideo";

export interface TmdbVideosResponse {
  id: number;
  results: TmdbVideo[];
}